requirements.txt
Authlib==1.6.5
blinker==1.9.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==2.23
python-dotenv==1.2.1
setuptools==80.9.0
Werkzeug==3.1.3
wheel==0.45.1
requests==2.32.5
